Feat: Heart-Redis-MySQL 기능 구현중

+++ BE/src/main/java/com/stella/stella/board/repository/HeartRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface HeartRepository extends JpaRepository<Heart, Long> {
-    int countByBoardBoardIndex(Long BoardIndex);
+    Long countByBoardBoardIndex(Long BoardIndex);
 
     List<Heart> findAllByMemberMemberIndex(Long MemberIndex);
